This is temporary!
I found a script called html2rst which extracts Docutils reST documents from HTML pages.
These files will have to be renamed, cleaned and validated to be put together.
The main benefit is that of a faster path to the documentation source conversion since performing
the conversion manually was a cumbersome task :-)


git-svn-id: http://svn.codehaus.org/izpack@1881 7d736ef5-cfd4-0310-9c9a-b52d5c14b761