Issue - IZPACK-69
Changed a few places in the installer and uninstaller where System.exit was being call to now use the Housekeeper.cleanup method instead.  This seems reasonable as users of Housekeeper really do not know when their listeners will be invoked and this change still invokes System.exit eventually with the same exit code as previous.


git-svn-id: http://svn.codehaus.org/izpack@2143 7d736ef5-cfd4-0310-9c9a-b52d5c14b761